requirements.txt
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator
pandas
python-multipart
openpyxl==3.1.2
pydantic>=2.10.0
fastapi>=0.115.0